Added class InducedVoltageResonator in impedances
Calculates the induced voltage of several resonators for arbitrary line
density. Takes a Beam and Slices object for the line density, a
Resonators object for the list of resonators, and an optional array of
time values where the induced voltage is to be calculated. If left out,
the induced voltage is calculated at the times of the line density.
Currently, it requires the all quality factors Q>0.5. Currently, only
works for single turn.