Fixed issue in matched_from_line_density
When a large number of impedance sources is used in 'matched_from_line_density', the computer could run out of RAM, because each impedance_source object calculated the impedance on a fine grid and stored it. This is unnecessary and caused memory overflow. 
Added the 'just_return' keyword argument for the 'imped_calc' and '(re)process' functions (default value is False). When True, the impedance is returned, rather than stored in the object. This is used in 'matched_from_line_density' to prevent memory overflow.