Arbitrary filling pattern for multi-bunch beam generation
The functions to generate multi-bunch beams now accept a list as 'bunch_spacing' arguments. This list contains the RF buckets of the bunches. That way, one can pass an arbitrary filling pattern. When 'bunch_spacing' is a number, the bunches are spaced 'bunch_spacing' RF buckets apart (this is also for backward compatibility)