Ivan's C-method & Option to shift beam & more unittests
Adopted Ivan's C-method for the quantum excitation. The method has a non-parallel call for the random numbers and a parallel call for the kicks. A 5% speed-up is obtained by combining the SR-kick and quantum excitation.
Included the option to shift the beam at the instantiation of class SynchrotronRadiation (default is True).